Cursor test to seperate file
